Create To find the top 10 countries for Rockbuster in terms of customer numbers
INNER JOIN was used to join the tables. It is cost effective. There are multiple joins in this query to connect the required tables. To obtain the desired information, the tables country and customer were joined, then address and customer, then city and address table and finally city table with country table.